src/dns: Refactor all `dns_*.rs` fines into a new module hierarchy
This one's a bit of a doozy, but in one fell swoop tries to shove all
`dns_*.rs`-named files into a `dns` module as they largely contain the
"guts" of processing DNS as a protocol.

There are a few unrelated changes mixed in, mostly converting `iopause`
as the one last remaining dependency outside the `dns` module itself.